// src/enums/ApplicationEnum.ts
export enum ApplicationEnum {
    HOOK_SYSTEM = "xotomicro_front_hook_system",
    EVENT_SERVICE = "xotomicro_front_event_service",
    ORDER_SERVICE = "xotomicro_front_order_service",
    PRODUCT_SERVICE = "xotomicro_front_product_service",
    USER_SERVICE = "xotomicro_front_user_service",
    AUTH_SERVICE = "xotomicro_front_auth_service",
}